Seperated out parsing into "src/parsing.c"
