+++ build.bash
@@ -1,3 +1,3 @@
 #! /bin/bash
-gcc src/main.c src/builtins.c -o clobber.out
+gcc src/main.c src/builtins.c src/parsing.c -o clobber.out
 ./clobber.out
